Moves smaller compsable hooks into shared helpers
